% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dateTrans.R
\name{dateTrans}
\alias{dateTrans}
\title{Formation of date.}
\usage{
dateTrans(date)
}
\arguments{
\item{date}{input a string(YYYYMMDD) showing the date}
}
\value{
the string denoting the date (YYYY-MM-DD)
}
\description{
Formation of date.
}
\keyword{basicFunc}

