% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dynamicUIStock.R
\name{dynamicUIStock}
\alias{dynamicUIStock}
\title{Refreshing stock data.}
\usage{
dynamicUIStock(symbol)
}
\arguments{
\item{symbol}{input a string in the double quotes showing the stock symbol}
}
\description{
Refreshing stock data.
}
\keyword{dynamicUIStock}

