% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getAllStockCur.R
\name{getAllStockCur}
\alias{getAllStockCur}
\title{get all the stock data of the latest day.}
\usage{
getAllStockCur()
}
\value{
data.frame contain all the data from the latest trading day

symbol Stock Symbol

code Stock Code

name Stock Name

trade the current trading price

pricechange the price change of the stock in RMB

changepercent the percent of the price change of the stock

buy buy price

sell sell price

settlement close price of the last day

open open price

high highest price

low lowest price

volume the volume

amount number of trading stocks * trading price

ticktime time

per Price-Earnings Ratio

pb Price/Book value

mktcap market capital

nmc market value

turnoverratio turnover ratio
}
\description{
get all the stock data of the latest day.
}
\examples{
getAllStockCur()
}
\keyword{getAllStockCur}

