% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Get Stock Data.}
\usage{
getData(symbol)
}
\arguments{
\item{symbol}{input a string show the stock symbol}
}
\value{
the stock data
}
\description{
Get Stock Data.
}
\keyword{basicFunc}

