% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getDayLabel.R
\name{getDayLabel}
\alias{getDayLabel}
\title{Get Day-based label.}
\usage{
getDayLabel(dayLabel)
}
\arguments{
\item{dayLabel}{input a string show the label}
}
\value{
the string denoting the data grouping
}
\description{
Get Day-based label.
}
\keyword{basicFunc}

