% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parsingDayprice.R
\name{parsingDayprice}
\alias{parsingDayprice}
\title{Get stock data of the specific pageNum.}
\usage{
parsingDayprice(pageNum = 1)
}
\arguments{
\item{pageNum}{the pageNum in the correlated website}
}
\description{
Get stock data of the specific pageNum.
}
\keyword{basicFunc}

